requirements.txt
asgiref==3.8.1
Django==5.2.1
django-extensions==4.1
pillow==11.2.1
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
djangorestframework==3.15.1
